\docType{package}
\name{MVPBT-package}
\alias{MVPBT-package}
\title{The 'MVPBT' package.}
\description{
Generalized Egger tests to detect publication bias in meta-analysis for diagnostic accuracy test.
}
\references{
Noma, H. (2020). 
Discussion of “Testing small study effects in multivariate meta-analysis” by Chuan Hong, Georgia Salanti, Sally Morton, Richard Riley, Haitao Chu, Stephen E. Kimmel, and Yong Chen. 
\emph{Biometrics} \strong{76}: 1255-1259.
\doi{10.1111/biom.13343}

Noma, H. (2022). 
MVPBT: R package for publication bias tests in meta-analysis of diagnostic accuracy studies. 
arXiv:2209.07270.
\doi{10.48550/arXiv.2209.07270}
}
